feat: auth middleware
Next.js middleware as of now requires it to be run on the Edge runtime. This runtime has limitations, including not being able to access crypto libraries needed for bcrypt, nor being able to run database operations. This kinda leaves us without a lot of options, if we want our middleware to perform auth checks.

So, unfortunately this leaves us with sub-optimal options. The best of which, IMO, is to run "middleware" directly from the routes themselves rather than implicitly using Next.js.

So remove the Next.js middleware.ts file, and replace it with our own middleware library. And use this library within our recommendations route.

For now this middleware confirms the auth cookie exists, is populated with a user's UUID, and matches an existing user's UUID. It returns a `Session` which contains only the `User` for now.