feat: add DB model Book and BookRecommendation (#31)
To better model the data we'll be dealing with, add Book and BookRecommendation to the database schema. Book allows us to share the data between the BookReview and BookRecommendation. For now, we're not storing the BookRecommendation to keep the changes to a minimum for this PR.

This change has ripple effects across the codebase. Update the code and tests to accommodate.