feat: nanoid as primary key
To avoid leaking information to the client about our data, switch to using nanoid's as the primary keys for the database tables. This allows us to remove the `uuid` field from User and use the `userId` instead.

There are lots of opinions on how to structure primary keys and how to expose public IDs for clients. In the end, this seemed like the simplest approach. We may regret this in the future? Famous last words? Probably nothing? Profit???