fix: useLoginError error url
This was a bit tricky to trace down, but noticed that when using the new protected context that the data was being reloaded on each page request. After a bit of digging, found out that the culprit was the fact that all the client API calls had a error catch to handle errors, and at it's core, `buildLoginErrorUrl` was being rebuilt each page request since `usePathname` was updating the path (this makes sense, you change a page and it requires a new path update). However, we shouldn't require all downstream uses to re-evaluate every page change, especially when we want to build state only once.

So, in place of that let's just use `window.location.pathname`, which is updated as the pages change but doesn't require React to re-build the function.